+++ spring-addons-starter-rest/src/main/java/com/c4_soft/springaddons/rest/synchronised/RestClientFactoryBean.java
@@ -2,6 +2,7 @@ package com.c4_soft.springaddons.rest.synchronised;
 
 import java.util.Optional;
 import org.springframework.beans.factory.FactoryBean;
+import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.lang.Nullable;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager;
 import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
@@ -16,8 +17,8 @@ public class RestClientFactoryBean implements FactoryBean<RestClient> {
   private SystemProxyProperties systemProxyProperties;
   private SpringAddonsRestProperties restProperties;
   private Optional<OAuth2AuthorizedClientManager> authorizedClientManager = Optional.empty();
-  private Optional<OAuth2AuthorizedClientRepository> authorizedClientRepository =
-      Optional.empty();
+  private Optional<OAuth2AuthorizedClientRepository> authorizedClientRepository = Optional.empty();
+  private Optional<ClientHttpRequestFactory> clientHttpRequestFactory;
 
   @Override
   @Nullable
@@ -28,6 +29,7 @@ public class RestClientFactoryBean implements FactoryBean<RestClient> {
     builderFactoryBean.setRestProperties(restProperties);
     builderFactoryBean.setAuthorizedClientManager(authorizedClientManager);
     builderFactoryBean.setAuthorizedClientRepository(authorizedClientRepository);
+    builderFactoryBean.setClientHttpRequestFactory(clientHttpRequestFactory);
     return Optional.ofNullable(builderFactoryBean.getObject()).map(RestClient.Builder::build)
         .orElse(null);
   }