+++ spring-addons-starter-rest/src/main/java/com/c4_soft/springaddons/rest/synchronised/SpringAddonsRestClientBeanDefinitionRegistryPostProcessor.java
@@ -63,6 +63,7 @@ public class SpringAddonsRestClientBeanDefinitionRegistryPostProcessor
           builder.addPropertyValue("restProperties", restProperties);
           builder.addAutowiredProperty("authorizedClientManager");
           builder.addAutowiredProperty("authorizedClientRepository");
+          builder.addAutowiredProperty("clientHttpRequestFactory");
           builder.addPropertyValue("clientId", e.getKey());
           registry.registerBeanDefinition(restProperties.getClientBeanName(e.getKey()),
               builder.getBeanDefinition());